# apps/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Postroom API",
    "version": "0.1.0",
    "description": "REST/JSON under /api for the Postroom webmail. Generated from the zod schemas that validate requests (PST-REQ-085); regenerate with `pnpm --filter @postroom/api openapi`.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "SenderPin": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "bucket": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "priority",
                  "people",
                  "newsletters",
                  "updates",
                  "receipts",
                  "notifications",
                  "junk"
                ]
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "address",
          "bucket"
        ],
        "additionalProperties": false
      },
      "SenderScreenResult": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "address": {
            "type": "string"
          },
          "decision": {
            "type": "string",
            "enum": [
              "allow",
              "block"
            ]
          },
          "clearedNewSender": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "ok",
          "address",
          "decision",
          "clearedNewSender"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "postroom_session"
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
...
    "/api/senders/{address}/pin": {
      "delete": {
        "operationId": "clearSenderPin",
        "tags": [
          "Senders"
        ],
        "summary": "Remove the bucket pin for this sender (a screen decision, if any, is unaffected).",
        "description": "Audited. Needs x-postroom-csrf: 1.",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 320,
              "pattern": "^[^\\s@]+@[^\\s@]+$"
            }
          },
          {
            "name": "x-postroom-csrf",
            "in": "header",
            "required": true,
            "description": "Must be 1.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok: true.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing CSRF header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getSenderPin",
        "tags": [
          "Senders"
        ],
        "summary": "The caller's pin for this sender address, if any (PST-REQ-105).",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 320,
              "pattern": "^[^\\s@]+@[^\\s@]+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pin (bucket is null when there is none).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SenderPin"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "setSenderPin",
        "tags": [
          "Senders"
        ],
        "summary": "Pin this sender to a bucket — overrides the classifier for their mail (PST-REQ-105).",
        "description": "A pin into Priority or People still requires the message to authenticate, exactly like the VIP rule. Audited. Needs x-postroom-csrf: 1.",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 320,
              "pattern": "^[^\\s@]+@[^\\s@]+$"
            }
          },
          {
            "name": "x-postroom-csrf",
            "in": "header",
            "required": true,
            "description": "Must be 1.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string",
                    "enum": [
                      "priority",
                      "people",
                      "newsletters",
                      "updates",
                      "receipts",
                      "notifications",
                      "junk"
                    ]
                  }
                },
                "required": [
                  "bucket"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The pin after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SenderPin"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing CSRF header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/senders/{address}/screen": {
      "post": {
        "operationId": "screenSender",
        "tags": [
          "Senders"
        ],
        "summary": "Answer this sender's new-sender badge: Allow or Block (PST-REQ-106).",
        "description": "Allow treats the sender as a known contact, so a directly-addressed message from them can reach Priority. Block routes their future mail to Junk. Either clears the $NewSender badge on their already-filed mail. Audited. Needs x-postroom-csrf: 1.",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 320,
              "pattern": "^[^\\s@]+@[^\\s@]+$"
            }
          },
          {
            "name": "x-postroom-csrf",
            "in": "header",
            "required": true,
            "description": "Must be 1.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "decision": {
                    "type": "string",
                    "enum": [
                      "allow",
                      "block"
                    ]
                  }
                },
                "required": [
                  "decision"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The screen decision after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SenderScreenResult"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Missing CSRF header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}